% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_text_check.R
\name{alt_text_check}
\alias{alt_text_check}
\title{Check alt text}
\usage{
alt_text_check(plot_name, alt_text_loc = c("console"), save_filepath)
}
\arguments{
\item{plot_name}{The plot to check for alt text}

\item{alt_text_loc}{Whether you want it displayed in the console, file, clipboard, or a combination}

\item{save_filepath}{If above set to file, where do you want to save it? #'}
}
\value{
Prints the alt text to console
}
\description{
Check a ggplot for the presence of alt text
}
\details{
#' Print alt text to console

Pass a plot to check whether it has alt text

Alt text is important to accessibility. This function checks for the presence of alt text in a ggplot object, and returns the contents of alt text when relevant.
}
\examples{
alt_text_check(dplyr::starwars |>
               ggplot2::ggplot(ggplot2::aes(x = height, y = mass)) +
               ggplot2::geom_point() +
               ggplot2::labs(alt = "This is some alt text"))

}
\seealso{
\code{\link[ggplot2]{get_alt_text}}
 \code{\link[cli]{cli_h1}}, \code{\link[cli]{cli_alert}}
 \code{\link[tools]{fileutils}}
}
