% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_text_suggest.R
\name{alt_text_suggest}
\alias{alt_text_suggest}
\title{Generate alt text automatically}
\usage{
alt_text_suggest(
  plot,
  prompt =
    "Please generate a descriptive alt text for a chart using the information below. Use the following formula: \\"Chart type (scatter plot, bar chart, line chart, etc.) + summary of the data (change in temperature, growth of sales, distribution of types of energy generation, etc.) + what it shows and why is it interesting (general trend, interesting outliers, etc.).",
  prompt_system =
    "You are a tool for helping ggplot2 users generate alt text in order to make their charts more accessible and SEO-friendly. You will be given a list of geoms used in the plot, the labels (like title, subtitle and annotations) and the data used to make the chart. Your job is to use that information to write alt text. You only return alt text, no other comments or feedback is necessary.",
  api_key = Sys.getenv("OPENAI_API_KEY"),
  model = "gpt-3.5-turbo-1106",
  temperature = 0,
  max_tokens = 2000,
  top_p = 1,
  frequency_penalty = 0,
  presence_penalty = 0
)
}
\arguments{
\item{prompt}{A question you want to ask GPT. Something like "Summarise this as a table".}

\item{prompt_system}{The system message helps set the behaviour of the AI. It can be something like "You are a helpful assistant."}

\item{api_key}{API key for GPT.}

\item{model}{Which model to use? If an API key is provided, it uses it to return the alt text to the console. Otherwise, it directs the user to ChatGPT.}

\item{plot_name}{The plot to check for alt text}
}
\value{
Prints the alt text to console
}
\description{
Generate alt text automatically
}
\details{
#' Generate alt text automatically

Pass a plot to genereate alt text for it using GPT

Generate alt text automatically.
}
\seealso{
\code{\link[ggplot2]{get_alt_text}}
}
