% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_bg.R
\docType{data}
\name{autocontrast}
\alias{autocontrast}
\title{Add the right colour to match the background}
\format{
An object of class \code{uneval} of length 1.
}
\usage{
autocontrast
}
\description{
Given a fill colour, return white or black so text is legible.
}
\examples{
library(ggplot2)
grid_data <- expand.grid(X = LETTERS[1:10], Y = paste0("var", seq(1, 10)))
grid_data$Z <- runif(100, 0, 5)

ggplot(grid_data, aes(X, Y, fill = Z)) +
  geom_tile() +
  geom_text(aes(label = Y, !!!autocontrast))
}
\keyword{datasets}
