% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_text.R
\name{check_alt_text}
\alias{check_alt_text}
\title{Print alt text to console}
\usage{
check_alt_text(plot_name, alt_text_loc = "console", save_filepath)
}
\arguments{
\item{plot_name}{The plot to check for alt text}

\item{alt_text_loc}{Whether you want it displayed in the console, saved to file, or both}

\item{save_filepath}{If above set to file, where do you want to save it?}
}
\description{
Pass a vector of colours and find out if they are colour-blind safe
}
\examples{
test_colours(c("red", "green", "darkgreen"), mode = "protan")

}
