% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_bg.R
\name{contrast_bg}
\alias{contrast_bg}
\title{Add the right colour to match the background}
\usage{
contrast_bg(colour, light_col = "white", dark_col = "black")
}
\arguments{
\item{colour}{The colour to check.}

\item{light_col}{The light output colour. Defaults to white.}

\item{dark_col}{The dark output colour. Defaults to black.}
}
\description{
Given a fill colour, return white or black so text is legible.
}
\examples{
contrast_bg("darkred") # returns white
contrast_bg("yellow") # returns black
}
