% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_check.R
\name{contrast_check}
\alias{contrast_check}
\title{Test text contrast for accessibility}
\usage{
contrast_check(col_1, col_2)
}
\arguments{
\item{col1}{The first colour}

\item{col2}{The second colour}
}
\description{
Pass two colours and print the result of a contrast check
}
\details{
Pass two hex codes and print the result of a contrast check

Pass two colours and print the result of a contrast check.
}
\examples{
check_threshold("black", "white")
check_threshold("yellow", "white")

}
