% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_check.R
\name{contrast_ratio}
\alias{contrast_ratio}
\title{#' @#' Test}
\usage{
contrast_ratio(colour, fill)
}
\arguments{
\item{col1}{The first colour}

\item{col2}{The second colour}
}
\value{
Returns a contrast ratio.
}
\description{
Pass two colours and receive a contrast ratio
}
\details{
Test text contrast for accessibility

Pass two hex codes and receive a contrast ratio

Pass two colours and receive a contrast ratio. Courtesy of coloratio.
}
\examples{
contrast_ratio("black", "white")
contrast_ratio("yellow", "white")

}
