% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_contrast.R
\name{test_contrast}
\alias{test_contrast}
\title{#' @#' Test}
\usage{
test_contrast(col_1, col_2, threshold = 4.5)
}
\arguments{
\item{col1}{The first colour}

\item{col2}{The second colour}
}
\value{
Returns a contrast ratio.
}
\description{
Pass two colours and receive a contrast ratio
}
\details{
#' Test text contrast for accessibility

Pass two colours and receive a contrast ratio

Pass two colours and receive a contrast ratio. Courtesy of coloratio.
}
\examples{
test_contrast("black", "white")
test_contrast("yellow", "white")

}
