% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plot.R
\name{test_plot}
\alias{test_plot}
\title{Test plot for colour blindness}
\usage{
test_plot(plot, test = c("cvd", "alt"))
}
\arguments{
\item{plot}{The plot you want to check}

\item{mode}{The tests you want to perform}
}
\description{
Find if a plot is colour-blind safe
}
\examples{
test_plot(plot_name, mode = "protan")

}
