% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\docType{data}
\name{autocontrast}
\alias{autocontrast}
\title{Add the right colour to match the background}
\format{
An object of class \code{uneval} of length 1.
}
\usage{
autocontrast
}
\description{
Given a fill colour, return white or black so text is legible.
}
\examples{
ggplot(df, aes(row, col, fill = value)) +
                 geom_raster() +
                 geom_text(aes(label = round(value, 2), !!!autocontrast))
}
\keyword{datasets}
