% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_text.R
\name{check_alt_text}
\alias{check_alt_text}
\title{Print alt text to console}
\usage{
check_alt_text(plot_name, alt_text_loc = "console", save_filepath)
}
\arguments{
\item{colours}{The colours you want to check}

\item{mode}{The tests you want to perform}
}
\description{
Pass a vector of colours and find out if they are colour-blind safe
}
\examples{
test_colours(c("red", "green", "darkgreen"), mode = "protan")

}
