% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_colours.R
\name{test_colourblind}
\alias{test_colourblind}
\title{Test colours for colour blindness}
\usage{
test_colourblind(
  colours,
  mode = c("deutan", "protan", "tritan"),
  smallestPercievableDistance = 9.2,
  ratioThres = 5
)
}
\arguments{
\item{colours}{The colours you want to check}

\item{mode}{The tests you want to perform}
}
\description{
Pass a vector of colours and find out if they are colour blind safe
}
\examples{
test_colourblind(c("red", "green", "darkgreen"), mode = "protan")

}
