% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plot.R
\name{test_plot}
\alias{test_plot}
\title{Test plot for accessibility}
\usage{
test_plot(plot, test = c("cvd", "alt"))
}
\arguments{
\item{plot}{The plot you want to check}

\item{test}{What test to run, Default: c("cvd", "alt")}
}
\value{
Prints potential accessibility issues to the console.
}
\description{
Find if a plot is colour-blind safe and return its alt text.
}
\details{
#' Test plot for colour blindness and alt text

Find if a plot is colour-blind safe and return its alt text

Test plot for accessibility.
}
\examples{
library(ggplot2)
starwars <- dplyr::starwars |>
  tidyr::drop_na(height, mass, gender)

starwars_plot <- starwars |>
  ggplot(aes(x = height, y = mass, colour = gender)) +
  geom_point(size = 4) +
  geom_label(aes(label = name), colour = "red", fill = "pink") +
  scale_colour_manual(values = c("red", "darkgreen"))

test_plot(starwars_plot)

}
\seealso{
\code{\link[ggplot2]{ggplot_build}}
 \code{\link[Ra11y]{test_colourblind}}, \code{\link[Ra11y]{check_alt_text}}
 \code{\link[cli]{cli_h1}}, \code{\link[cli]{cli_alert}}, \code{\link[cli]{cli_ul}}, \code{\link[cli]{cli_text}}
}
